requirement.txt
streamlit
rembg
Pillow
opencv-python
mediapipe
icrawler
numpy
